requirements.txt
https://www.djangoproject.com/download/1.7.b4/tarball/
distribute
psycopg2
djangorestframework==2.3.13
django-allauth
